move states to App.jsx level
