+++ docs/components-relationship.drawio
@@ -1,13 +1,13 @@
-<mxfile host="app.diagrams.net" modified="2023-09-01T07:22:54.189Z" agent="Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/116.0.0.0 Safari/537.36" etag="xkUqOm08kyW5Tk59wadr" version="21.7.0" type="device">
+<mxfile host="app.diagrams.net" modified="2023-09-01T12:41:14.596Z" agent="Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/116.0.0.0 Safari/537.36" etag="JcjtqOd556O6b3nGhG5H" version="21.7.0" type="device">
   <diagram name="Page-1" id="_32pdijGtaBiMeI2xC9d">
-    <mxGraphModel dx="1647" dy="853" grid="1" gridSize="10" guides="1" tooltips="1" connect="1" arrows="1" fold="1" page="1" pageScale="1" pageWidth="1169" pageHeight="827" math="0" shadow="1">
+    <mxGraphModel dx="1101" dy="636" grid="1" gridSize="10" guides="1" tooltips="1" connect="1" arrows="1" fold="1" page="1" pageScale="1" pageWidth="1169" pageHeight="827" math="0" shadow="1">
       <root>
         <mxCell id="0" />
         <mxCell id="1" parent="0" />
-        <mxCell id="wLTWeM-jETooPy9-A8mS-15" value="quote.author" style="edgeStyle=orthogonalEdgeStyle;rounded=1;hachureGap=4;orthogonalLoop=1;jettySize=auto;html=1;exitX=0;exitY=0.75;exitDx=0;exitDy=0;fontFamily=Helvetica;fontSize=16;fontColor=default;entryX=0.75;entryY=0;entryDx=0;entryDy=0;fillColor=#0050ef;strokeColor=#001DBC;shape=flexArrow;" edge="1" parent="1" source="wLTWeM-jETooPy9-A8mS-1" target="wLTWeM-jETooPy9-A8mS-2">
+        <mxCell id="wLTWeM-jETooPy9-A8mS-15" value="&lt;ul style=&quot;font-size: 14px;&quot;&gt;&lt;li style=&quot;text-align: left;&quot;&gt;`quote.author`&lt;/li&gt;&lt;li style=&quot;text-align: left;&quot;&gt;`textVisible` state &lt;br&gt;(for fading)&lt;/li&gt;&lt;/ul&gt;" style="edgeStyle=orthogonalEdgeStyle;rounded=1;hachureGap=4;orthogonalLoop=1;jettySize=auto;html=1;exitX=0;exitY=0.75;exitDx=0;exitDy=0;fontFamily=Helvetica;fontSize=16;fontColor=default;entryX=0.75;entryY=0;entryDx=0;entryDy=0;fillColor=#0050ef;strokeColor=#001DBC;shape=flexArrow;" edge="1" parent="1" source="wLTWeM-jETooPy9-A8mS-1" target="wLTWeM-jETooPy9-A8mS-2">
           <mxGeometry relative="1" as="geometry" />
         </mxCell>
-        <mxCell id="wLTWeM-jETooPy9-A8mS-16" value="quote" style="edgeStyle=orthogonalEdgeStyle;rounded=1;hachureGap=4;orthogonalLoop=1;jettySize=auto;html=1;exitX=0.75;exitY=1;exitDx=0;exitDy=0;fontFamily=Helvetica;fontSize=16;fontColor=default;entryX=0.75;entryY=0;entryDx=0;entryDy=0;fillColor=#0050ef;strokeColor=#001DBC;shape=flexArrow;" edge="1" parent="1" source="wLTWeM-jETooPy9-A8mS-1" target="wLTWeM-jETooPy9-A8mS-4">
+        <mxCell id="wLTWeM-jETooPy9-A8mS-16" value="&lt;ul style=&quot;font-size: 14px;&quot;&gt;&lt;li&gt;`quote`&lt;/li&gt;&lt;li&gt;`textVisible` state&lt;br&gt;(for fading)&lt;/li&gt;&lt;/ul&gt;" style="edgeStyle=orthogonalEdgeStyle;rounded=1;hachureGap=4;orthogonalLoop=1;jettySize=auto;html=1;exitX=0.75;exitY=1;exitDx=0;exitDy=0;fontFamily=Helvetica;fontSize=16;fontColor=default;entryX=0.75;entryY=0;entryDx=0;entryDy=0;fillColor=#0050ef;strokeColor=#001DBC;shape=flexArrow;align=left;width=10;" edge="1" parent="1" source="wLTWeM-jETooPy9-A8mS-1" target="wLTWeM-jETooPy9-A8mS-4">
           <mxGeometry relative="1" as="geometry">
             <mxPoint x="540" y="414" as="sourcePoint" />
           </mxGeometry>
