+++ src/App.jsx
@@ -1,13 +1,41 @@
+import { useState } from "react";
 import QuoteBox from "./components/QuoteBox";
 import BackgroundImage from "./components/BackgroundImage";
+import quotes from "../data/quotes.json";
 import styles from "./styles/App.module.css";
 
-function App() {
+export default function App() {
+  const getRandomQuoteIndex = () => Math.floor(Math.random() * quotes.length);
+
+  const [quoteIndex, setQuoteIndex] = useState(getRandomQuoteIndex());
+  const [textVisible, setTextVisible] = useState(true);
+
+  const handleNewQuoteButtonClick = () => {
+    const updateQuoteIndex = () => {
+      let newQuoteIndex = getRandomQuoteIndex();
+      while (newQuoteIndex === quoteIndex) {
+        newQuoteIndex = getRandomQuoteIndex();
+      }
+      setQuoteIndex(newQuoteIndex);
+    };
+
+    // fades text (quote and author) out
+    setTextVisible(false);
+    setTimeout(() => {
+      updateQuoteIndex();
+      // fades text back in
+      setTextVisible(true);
+    }, 550);
+  };
+
   return (
     <div className={styles.background}>
-      <QuoteBox className={styles.quoteBox} />
+      <QuoteBox
+        className={styles.quoteBox}
+        textVisible={textVisible}
+        quoteIndex={quoteIndex}
+        handleNewQuoteButtonClick={handleNewQuoteButtonClick}
+      />
     </div>
   );
 }
-
-export default App;
